Reduce code duplication by moving partition_lomuto_branchless_simple into quicksort module
